# skills/bkd/references/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BKD API",
    "description": "Kanban board for managing AI coding agents. Issues are assigned to CLI-based AI engines (Claude Code, Codex, Gemini CLI) that execute autonomously.",
    "version": "dev",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "CronJob": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "cron": {
            "type": "string"
          },
          "taskType": {
            "type": "string"
          },
          "taskConfig": {
            "type": "object",
            "additionalProperties": {}
          },
          "enabled": {
            "type": "boolean"
          },
          "lastRun": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "cron",
          "taskType",
          "taskConfig",
          "enabled",
          "lastRun",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCronJob": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "cron": {
            "type": "string",
            "minLength": 1
          },
          "action": {
            "type": "string",
            "minLength": 1
          },
          "config": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "name",
          "cron",
          "action"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/cron": {
      "get": {
        "tags": [
          "Cron"
        ],
        "summary": "List cron jobs",
        "operationId": "listCronJobs",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "only"
              ]
            },
            "required": false,
            "name": "deleted",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Cron jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CronJob"
                          }
                        },
                        {
                          "type": "object",
                          "properties": {
                            "jobs": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/CronJob"
                              }
                            },
                            "hasMore": {
                              "type": "boolean"
                            },
                            "nextCursor": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "jobs",
                            "hasMore",
                            "nextCursor"
                          ]
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cron"
        ],
        "summary": "Create cron job",
        "operationId": "createCronJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCronJob"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created job",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CronJob"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Job name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cron/{jobId}": {
      "delete": {
        "tags": [
          "Cron"
        ],
        "summary": "Soft-delete cron job",
        "operationId": "deleteCronJob",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "jobId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "deleted": {
                          "type": "boolean"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "deleted",
                        "name"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cron/{jobId}/logs": {
      "get": {
        "tags": [
          "Cron"
        ],
        "summary": "Get logs for a cron job",
        "operationId": "getCronJobLogs",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "jobId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "success",
                "failed",
                "running"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Job logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobName": {
                          "type": "string"
                        },
                        "logs": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "startedAt": {
                                "type": "string"
                              },
                              "finishedAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "durationMs": {
                                "type": [
                                  "integer",
                                  "null"
                                ]
                              },
                              "status": {
                                "type": "string"
                              },
                              "result": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "error": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "startedAt",
                              "finishedAt",
                              "durationMs",
                              "status",
                              "result",
                              "error"
                            ]
                          }
                        },
                        "hasMore": {
                          "type": "boolean"
                        },
                        "nextCursor": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "jobName",
                        "logs",
                        "hasMore",
                        "nextCursor"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cron/{jobId}/trigger": {
      "post": {
        "tags": [
          "Cron"
        ],
        "summary": "Manually trigger a cron job",
        "operationId": "triggerCronJob",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "jobId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Triggered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "triggered": {
                          "type": "boolean"
                        },
                        "name": {
                          "type": "string"
                        },
                        "log": {
                          "type": [
                            "object",
                            "null"
                          ],
                          "properties": {
                            "status": {
                              "type": "string"
                            },
                            "durationMs": {
                              "type": [
                                "integer",
                                "null"
                              ]
                            },
                            "result": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "error": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "status",
                            "durationMs",
                            "result",
                            "error"
                          ]
                        }
                      },
                      "required": [
                        "triggered",
                        "name",
                        "log"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Job already running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cron/{jobId}/pause": {
      "post": {
        "tags": [
          "Cron"
        ],
        "summary": "Pause a cron job",
        "operationId": "pauseCronJob",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "jobId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Paused",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "paused": {
                          "type": "boolean"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "paused",
                        "name"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/cron/{jobId}/resume": {
      "post": {
        "tags": [
          "Cron"
        ],
        "summary": "Resume a paused cron job",
        "operationId": "resumeCronJob",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "jobId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Resumed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "resumed": {
                          "type": "boolean"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "resumed",
                        "name"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
}